Web Deployment Testing
Removed abc.php and ajcalendar.php as the former was a test, and the
latter is obsolete now that we're no longer using ajax for the calendar.

Worked with index.php, config.php and database.php to deploy remotely.

EXT_Controller and Dashboard.php are renamed for web deployment.
Web-server PHP isn't responding to class linkage with lowercase
filenames and UpperCamelCase class definitions.  Need to use single
Uppercaseclasses from now on and proceed to rename all classes under
that scheme.